# docs/swagger.yaml
{
  "swagger": "2.0",
  "info": {
    "contact": {}
  },
  "paths": {
...
    "/messages/get": {
      "get": {
        "description": "Fetches all messages exchanged with a specific recipient.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Messages"
        ],
        "parameters": [
          {
            "type": "string",
            "description": "Recipient ID",
            "name": "reciever",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.GenericResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/models.GenericResponse"
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "schema": {
              "$ref": "#/definitions/models.GenericResponse"
            }
          },
          "406": {
            "description": "Not Acceptable",
            "schema": {
              "$ref": "#/definitions/models.GenericResponse"
            }
          }
        }
      }
    },
    "/messages/sent": {
      "post": {
        "description": "Sends a new message from the authorized user to the recipient.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Messages"
        ],
        "parameters": [
          {
            "description": "Message payload",
            "name": "requestBody",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.Message"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.GenericResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/models.GenericResponse"
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "schema": {
              "$ref": "#/definitions/models.GenericResponse"
            }
          }
        }
      }
    },
    "/user/deleteUser": {
      "delete": {
        "description": "Deletes a user account by the specified username.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "User Management"
        ],
        "parameters": [
          {
            "type": "string",
            "description": "Username of the user to delete",
            "name": "username",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.GenericResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/models.GenericResponse"
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "schema": {
              "$ref": "#/definitions/models.GenericResponse"
            }
          },
          "406": {
            "description": "Not Acceptable",
            "schema": {
              "$ref": "#/definitions/models.GenericResponse"
            }
          }
        }
      }
    },
    "/user/fetchUser": {
      "get": {
        "description": "Handles requests to retrieve user details using a username provided in the query parameters.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Users"
        ],
        "summary": "Fetch user details",
        "parameters": [
          {
            "type": "string",
            "description": "Username of the user to fetch",
            "name": "username",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.GenericResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/models.GenericResponse"
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "schema": {
              "$ref": "#/definitions/models.GenericResponse"
            }
          },
          "406": {
            "description": "Not Acceptable",
            "schema": {
              "$ref": "#/definitions/models.GenericResponse"
            }
          }
        }
      }
    },
    "/user/updateUserAndProfile": {
      "post": {
        "description": "Updates the details of a user based on the provided username and JSON body payload.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Users"
        ],
        "summary": "Update user and profile details",
        "parameters": [
          {
            "type": "string",
            "description": "Username of the user to update",
            "name": "username",
            "in": "query",
            "required": true
          },
          {
            "description": "JSON payload containing updated user details",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.UpdateUserAndProfile"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.GenericResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/models.GenericResponse"
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "schema": {
              "$ref": "#/definitions/models.GenericResponse"
            }
          },
          "406": {
            "description": "Not Acceptable",
            "schema": {
              "$ref": "#/definitions/models.GenericResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "models.GenericResponse": {
      "type": "object",
      "properties": {
        "data": {},
        "message": {
          "type": "string"
        },
        "status": {
          "type": "boolean"
        }
      }
    },
...
    "models.Message": {
      "type": "object",
      "properties": {
        "chat_id": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "media_url": {
          "type": "string"
        },
        "message_id": {
          "type": "string"
        },
        "recipient_id": {
          "type": "string"
        },
        "sender_id": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "timestamp": {
          "type": "string"
        }
      }
    },
    "models.Profile": {
      "type": "object",
      "properties": {
        "achievements": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "bio": {
          "type": "string"
        },
        "contact_preferences": {
          "type": "object",
          "additionalProperties": {
            "type": "boolean"
          }
        },
        "cover_photo_url": {
          "type": "string"
        },
        "education": {
          "type": "string"
        },
        "gender": {
          "type": "string"
        },
        "interests": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "is_profile_public": {
          "type": "boolean"
        },
        "occupation": {
          "type": "string"
        },
        "phone_number": {
          "type": "string"
        },
        "profile_completeness": {
          "type": "integer"
        },
        "social_links": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        }
      }
    },
    "models.Role": {
      "type": "string",
      "enum": [
        "ADMIN",
        "CLIENT"
      ],
      "x-enum-varnames": [
        "Admin",
        "Client"
      ]
    },
    "models.UpdateUserAndProfile": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string"
        },
        "avatar_url": {
          "type": "string"
        },
        "date_of_birth": {
          "type": "string"
        },
        "first_name": {
          "type": "string"
        },
        "last_name": {
          "type": "string"
        },
        "last_seen": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "profile": {
          "$ref": "#/definitions/models.Profile"
        },
        "role": {
          "$ref": "#/definitions/models.Role"
        },
        "status_message": {
          "type": "string"
        }
      }
    },
...
  }
}